+++ src/app/components/inviteAccept/inviteAcceptController.js
@@ -40,8 +40,8 @@ angular.module('liveopsConfigPanel')
         user.password = $scope.newPassword;
 
         TenantUser.update({
-          tenantId: $stateParams.tenantId,
-          id: $stateParams.userId,
+          tenantId: $stateParams.tenantId || $stateParams.tenantid,
+          id: $stateParams.userId || $stateParams.userid,
           status: 'accepted'
         }, $scope.acceptSuccess, $scope.acceptFailure);
       };
@@ -57,7 +57,7 @@ angular.module('liveopsConfigPanel')
         return AuthService.login($scope.user.email, $scope.newPassword).then(function() {
           if (UserPermissions.hasPermissionInList(['PLATFORM_MANAGE_ALL_TENANTS_ENROLLMENT', 'VIEW_ALL_USERS', 'MANAGE_ALL_USER_EXTENSIONS', 'MANAGE_ALL_GROUP_USERS', 'MANAGE_ALL_USER_SKILLS', 'MANAGE_ALL_USER_LOCATIONS', 'MANAGE_TENANT_ENROLLMENT'])) {
             $state.transitionTo('content.management.users', {
-              id: $stateParams.userId,
+              id: $stateParams.userid,
               messageKey: 'invite.accept.autologin.success'
             });
           } else {