+++ src/app/components/management/users/userTableConfig.js
@@ -48,6 +48,12 @@ angular.module('liveopsConfigPanel')
               'display': $translate.instant('value.email')
             },
             'name': '$user.$original.email'
+          }, {
+            'header': {
+              'display': $translate.instant('details.externalId')
+            },
+            'name': '$user.$original.externalId',
+            'checked': false
           }],
           'orderBy': '$user.$original.lastName',
           'sref': 'content.management.users',
@@ -175,8 +181,8 @@ angular.module('liveopsConfigPanel')
             'displayPath': 'display',
             'options': userStatuses()
           },
-          'name': '$original.$user.status',
-          'lookup': '$original:$user:status',
+          'name': '$original.status',
+          'lookup': '$original:status',
           'id': 'user-status-table-column',
           'transclude': true,
           'checked': false